Fix build/styling errors
