+++ packages/js/logging-server/tsconfig.json
@@ -16,7 +16,8 @@
     "sourceMap": true,
     "target": "es5",
     "resolveJsonModule": true,
-    "strictNullChecks": true
+    "strictNullChecks": true,
+    "outDir": "build"
   },
   "include": ["./src/**/*"]
 }
